@@ -1,5 +1,5 @@
 name: smartpay
-description: A new Flutter project.
+description: Smart Pay App.
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -16,10 +16,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.0+1
+version: 1.0.0+2
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '^3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -35,22 +35,23 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  http: ^0.13.6
-  google_fonts: ^4.0.4
-  flutter_dotenv: ^5.0.2
-  flutter_riverpod: ^2.3.6
-  transparent_image: ^2.0.1
+  http:
+  google_fonts:
+  flutter_dotenv:
+  flutter_riverpod:
+  transparent_image:
   intl:
   syncfusion_flutter_calendar:
-  flutter_localization:
   syncfusion_localizations:
   syncfusion_flutter_pdfviewer:
-  permission_handler: ^10.3.0
-  path_provider: ^2.0.15
-  camera: ^0.10.5+2
-  file_picker: ^5.3.2
-  geolocator: ^10.0.0
-  qr_code_scanner: ^1.0.1
+  permission_handler:
+  path_provider:
+  camera:
+  file_picker: #^5.3.2
+  geolocator:
+  qr_code_scanner:
+  flutter_localizations:
+    sdk: flutter
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -60,7 +61,13 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints:
+
+  flutter_launcher_icons:
+
+flutter_icons:
+  android: true
+  image_path: "assets/images/logo.jpeg"
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
