@@ -51,16 +51,15 @@ class FileStorage {
       // If not we will ask for permission first
       await Permission.storage.request();
     }
-    Directory _directory = Directory("");
+    Directory directory = Directory("");
     if (Platform.isAndroid) {
       // Redirects it to download folder in android
-      _directory = Directory("/storage/emulated/0/Download");
+      directory = Directory("/storage/emulated/0/Download");
     } else {
-      _directory = await getApplicationDocumentsDirectory();
+      directory = await getApplicationDocumentsDirectory();
     }
 
-    final exPath = _directory.path;
-    print("Saved Path: $exPath");
+    final exPath = directory.path;
     await Directory(exPath).create(recursive: true);
     return exPath;
   }
@@ -78,8 +77,6 @@ class FileStorage {
     // Create a file for the path of
     // device and file name with extension
     File file = File('$path/$name');
-    ;
-    print("Save file");
 
     // Write the data in the file you have created
     return file.writeAsString(bytes);
